+++ pkg/util/utils_test.go
@@ -22,88 +22,7 @@ import (
 	"strings"
 	"sync"
 	"testing"
-
-	"github.com/pkg/errors"
-	"k8s.io/minikube/pkg/util/retry"
 )
-
-// Returns a function that will return n errors, then return successfully forever.
-func errorGenerator(n int, retryable bool) func() error {
-	errorCount := 0
-	return func() (err error) {
-		if errorCount < n {
-			errorCount++
-			e := errors.New("Error")
-			if retryable {
-				return &retry.RetriableError{Err: e}
-			}
-			return e
-
-		}
-
-		return nil
-	}
-}
-
-func TestErrorGenerator(t *testing.T) {
-	errors := 3
-	f := errorGenerator(errors, false)
-	for i := 0; i < errors-1; i++ {
-		if err := f(); err == nil {
-			t.Fatalf("Error should have been thrown at iteration %v", i)
-		}
-	}
-	if err := f(); err == nil {
-		t.Fatalf("Error should not have been thrown this call!")
-	}
-}
-
-func TestRetry(t *testing.T) {
-	f := errorGenerator(4, true)
-	if err := Retry(5, f); err != nil {
-		t.Fatalf("Error should not have been raised by retry.")
-	}
-
-	f = errorGenerator(5, true)
-	if err := Retry(4, f); err == nil {
-		t.Fatalf("Error should have been raised by retry.")
-	}
-}
-
-func TestRetryNotRetriableError(t *testing.T) {
-	f := errorGenerator(4, false)
-	if err := Retry(5, f); err == nil {
-		t.Fatalf("Error should have been raised by retry.")
-	}
-
-	f = errorGenerator(5, false)
-	if err := Retry(4, f); err == nil {
-		t.Fatalf("Error should have been raised by retry.")
-	}
-	f = errorGenerator(0, false)
-	if err := Retry(5, f); err != nil {
-		t.Fatalf("Error should not have been raised by retry.")
-	}
-}
-
-func TestMultiError(t *testing.T) {
-	m := MultiError{}
-
-	m.Collect(errors.New("Error 1"))
-	m.Collect(errors.New("Error 2"))
-
-	err := m.ToError()
-	expected := `Error 1
-Error 2`
-	if err.Error() != expected {
-		t.Fatalf("%s != %s", err.Error(), expected)
-	}
-
-	m = MultiError{}
-	if err := m.ToError(); err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-}
 
 func TestGetBinaryDownloadURL(t *testing.T) {
 	testData := []struct {
